feat: update e-commerce dataset pipeline
