--- datasets/thelook_ecommerce/.gitignore
@@ -1,4 +1,0 @@
-image_history
-test.py
-_images/run_thelook_kub/*.json
-_images/run_thelook_kub/helper
