--- datasets/thelook_ecommerce/_images/run_thelook_kub/requirements.txt
@@ -1,46 +1,0 @@
-aiohttp==3.8.1
-aiosignal==1.2.0
-async-timeout==4.0.2
-attrs==21.4.0
-cachetools==4.2.2
-certifi==2021.5.30
-charset-normalizer==2.0.4
-decorator==5.1.1
-faker==8.12.1
-fsspec==2022.1.0
-gcsfs==2022.1.0
-google-api-core==2.0.1
-google-api-python-client==2.20.0
-google-auth==2.0.2
-google-auth-httplib2==0.1.0
-google-auth-oauthlib==0.4.6
-google-cloud-core==2.2.2
-google-cloud-storage==2.1.0
-google-crc32c==1.3.0
-google-resumable-media==2.1.0
-googleapis-common-protos==1.53.0
-gspread==4.0.1
-httplib2==0.19.1
-idna==3.2
-install==1.3.4
-multidict==6.0.2
-numpy==1.21.2
-oauth2client==4.1.3
-oauthlib==3.1.1
-pandas==1.3.2
-protobuf==3.17.3
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
-pyparsing==2.4.7
-python-dateutil==2.8.2
-pytz==2021.1
-requests==2.26.0
-requests-oauthlib==1.3.0
-rsa==4.7.2
-six==1.16.0
-text-unidecode==1.3
-timedelta==2020.12.3
-typing==3.7.4.3
-uritemplate==3.0.1
-urllib3==1.26.6
-yarl==1.7.2
