+++ datasets/thelook_ecommerce/infra/thelook_ecommerce_pipeline.tf
@@ -1,0 +1,39 @@
+/**
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+
+resource "google_bigquery_table" "thelook_ecommerce_thelook_ecommerce" {
+  project    = var.project_id
+  dataset_id = "thelook_ecommerce"
+  table_id   = "thelook_ecommerce"
+
+  description = "The Look E-Commerce Dataset"
+
+
+
+
+  depends_on = [
+    google_bigquery_dataset.thelook_ecommerce
+  ]
+}
+
+output "bigquery_table-thelook_ecommerce_thelook_ecommerce-table_id" {
+  value = google_bigquery_table.thelook_ecommerce_thelook_ecommerce.table_id
+}
+
+output "bigquery_table-thelook_ecommerce_thelook_ecommerce-id" {
+  value = google_bigquery_table.thelook_ecommerce_thelook_ecommerce.id
+}
